+++ base/deprecated.jl
@@ -1436,6 +1436,9 @@ function slicedim(A::AbstractVector, d::Integer, i::Number)
     end
 end
 
+# Issue #25786
+@deprecate_binding DevNull devnull
+
 # PR 25062
 @deprecate(link_pipe(pipe; julia_only_read = true, julia_only_write = true),
            link_pipe!(pipe, reader_supports_async = julia_only_read, writer_supports_async = julia_only_write),